2018.02.08
- refactor server
- add “servlet” API layer
- speed up file and CGI requests
- modify AltUI scene logging
